feat(Vault): Adds close vault feauture
In order for the bounty pool value on a period to be calculated, we need to define a point in time
where no more deposits are accepted. This feature adds the `close` function, which closes the vault,
and prevents further deposits from being made for that period. This feature also prevents transfers
from being made before the vault is closed. In summary, deposits can only be made when the vault is
open. Withdrawing can only be made when the vault is closed.

BREAKING CHANGE: The `vault.close(args)`function needs to be called before withdrawals can be done